fix(file organization): add viewmodel and clean coordinator
- add ShareView and ShareViewModel
- Remove multiple views in case in MainCoordinator
