+++ swifty-protein/Coordinators/MainCoordinator.swift
@@ -8,7 +8,6 @@
 import SwiftUI
 
 struct MainCoordinator: View {
-    @State private var capturedView: UIView?
     @State private var path: [PushableView] = []
     @State private var introDidComplete: Bool = false
     
@@ -50,48 +49,19 @@ internal extension MainCoordinator {
                 path.append(.detail(ligand))
             }
         case let .detail(ligand):
-            ShareableDetailView(searchText: ligand.id, onError: { path.removeLast() }) { view in
-                self.capturedView = view
+            ShareView(seachText: ligand.id, onError: { path.removeLast() }, viewModel: .init())
+        }
+    }
+    
+    @ViewBuilder
+    var contentView: some View {
+        if introDidComplete {
+            LoginView(viewModel: .init()) {
+                path.append(.search)
             }
-                .navigationTitle(ligand.id)
-                .navigationBarTitleDisplayMode(NavigationBarItem.TitleDisplayMode.large)
-                                Button(action: {
-                if let view = capturedView {
-                    if let image = captureScreen(view: view) {
-                        shareImage(image)
-                    }
-                }
-            }, label: {
-                Image(systemName: "square.and.arrow.up")
-                    .padding()
-            })
-                                }
-                                }
-                                func captureScreen(view: UIView) -> UIImage? {
-                let renderer = UIGraphicsImageRenderer(size: view.bounds.size)
-                return renderer.image { _ in
-                    view.drawHierarchy(in: view.bounds, afterScreenUpdates: true)
-                }
-            }
-                                
-                                func shareImage(_ image: UIImage) {
-                let av = UIActivityViewController(activityItems: [image], applicationActivities: nil)
-                
-                if let windowScene = UIApplication.shared.connectedScenes.first as? UIWindowScene,
-                   let rootViewController = windowScene.windows.first?.rootViewController {
-                    rootViewController.present(av, animated: true, completion: nil)
-                }
-            }
-                                
-                                @ViewBuilder
-                                var contentView: some View {
-                if introDidComplete {
-                    LoginView(viewModel: .init()) {
-                        path.append(.search)
-                    }
-                } else {
-                    LaunchScreenView()
-                        .transition(.opacity)
-                }
-            }
-                                }
+        } else {
+            LaunchScreenView()
+                .transition(.opacity)
+        }
+    }
+}
