+++ swifty-protein/Views/ShareView/ShareViewModel.swift
@@ -1,0 +1,27 @@
+//
+//  ShareViewModel.swift
+//  swifty-protein
+//
+//  Created by Léa Lescure on 12/10/2023.
+//
+
+import Foundation
+import UIKit
+
+final class ShareViewModel: ObservableObject {
+    func captureScreen(view: UIView) -> UIImage? {
+        let renderer = UIGraphicsImageRenderer(size: view.bounds.size)
+        return renderer.image { _ in
+            view.drawHierarchy(in: view.bounds, afterScreenUpdates: true)
+        }
+    }
+    
+    func shareImage(_ image: UIImage) {
+        let av = UIActivityViewController(activityItems: [image], applicationActivities: nil)
+        
+        if let windowScene = UIApplication.shared.connectedScenes.first as? UIWindowScene,
+           let rootViewController = windowScene.windows.first?.rootViewController {
+            rootViewController.present(av, animated: true, completion: nil)
+        }
+    }
+}
