+++ swifty-protein/Views/ShareView/ShareableDetailView.swift
